// packages/sdk/src/generated/chain.ts
// Generated by scripts/sync-abi.mjs from packages/contracts. Do not edit by hand.

export const chainConfig = {
  "chainId": 4663,
  "name": "Robinhood Chain",
  "rpc": "https://rpc.mainnet.chain.robinhood.com",
  "explorer": "https://robinhoodchain.blockscout.com",
  "usdg": "0x5fc5360D0400a0Fd4f2af552ADD042D716F1d168",
  "weth": "0x0Bd7D308f8E1639FAb988df18A8011f41EAcAD73",
  "uniswapV4": {
    "poolManager": "0x8366a39CC670B4001A1121B8F6A443A643e40951",
    "quoter": "0x8dc178efb8111bb0973dd9d722ebeff267c98f94",
    "positionManager": "0x58daec3116aae6d93017baaea7749052e8a04fa7",
    "stateView": "0xf3334192d15450cdd385c8b70e03f9a6bd9e673b",
    "universalRouter": "0x8876789976decbfcbbbe364623c63652db8c0904",
    "permit2": "0x000000000022D473030F116dDEE9F6B43aC78BA3"
  },
  "pons": {
    "launchFactory": "0x7eD598BcEf8bd9Edd8C97A195C6d13f40801EC7e",
    "memeHook": "0xE5e702641Ea86F4ae6cC3cDaeD2B886f976Be044",
    "graduatedPool": {
      "fee": 0,
      "tickSpacing": 200
    }
  },
  "x402": {
    "network": "eip155:4663",
    "asset": {
      "address": "0x5fc5360D0400a0Fd4f2af552ADD042D716F1d168",
      "decimals": 6,
      "eip712": {
        "name": "Global Dollar",
        "version": "1"
      }
    },
    "facilitators": [
      {
        "name": "Naven",
        "url": "https://facilitator.naven.network"
      },
      {
        "name": "Primer",
        "url": "https://x402.primer.systems"
      }
    ]
  }
} as const;