// packages/sdk/src/generated/deployments.ts
// Generated by scripts/sync-abi.mjs from packages/contracts. Do not edit by hand.

export const deployments = {
  "chainId": 4663,
  "contracts": {
    "BudgetAccountFactory": {
      "address": "0x68b5bb42fec83db9582758bbcb1fc43f748970d6",
      "txHash": "0x74a7e1b436d1014d295a34801b452b38b82c97e1ebf34770e62b90f8703c623e",
      "block": 54087837,
      "deployedAt": "2026-09-04T07:29:30Z",
      "constructorArgs": "0x0000000000000000000000005fc5360d0400a0fd4f2af552add042d716f1d168000000000000000000000000fceb8905e316d383cd90aa1ab04ab1650611445b",
      "verified": "sourcify",
      "implementation": "0x1Cc03Bab3B6d9a5059cEC6a1BEFa86982D492C23"
    },
    "BurnVault": {
      "address": "0x80d9bc2412853030f259ea7056654888b2b0d768",
      "txHash": "0x7c3c329698c1fe51d333ec26d13fe7eedf80afdbc8c9d10f44da5c0ef3399281",
      "block": 55666987,
      "deployedAt": "2026-09-06T03:46:22Z",
      "constructorArgs": "0x0000000000000000000000008366a39cc670b4001a1121b8f6a443a643e40951000000000000000000000000caf681a66d020601342297493863e78c959e5cb20000000000000000000000005fc5360d0400a0fd4f2af552add042d716f1d1680000000000000000000000000bd7d308f8e1639fab988df18a8011f41eacad73000000000000000000000000fceb8905e316d383cd90aa1ab04ab1650611445b",
      "verified": "sourcify"
    },
    "CreatorRegistry": {
      "address": "0x5d483aec0735d550d09018a2e89c49c190962deb",
      "txHash": "0xf1e188b3d42632f9d2d754b22ac6c483447ec9f6f9fa096d928cbdfae9b14bbe",
      "block": 54087895,
      "deployedAt": "2026-09-04T07:29:30Z",
      "constructorArgs": "0x0000000000000000000000005fc5360d0400a0fd4f2af552add042d716f1d1680000000000000000000000007cf3852a9da0064cb22ca69f5726adb16dfa34cb000000000000000000000000fceb8905e316d383cd90aa1ab04ab1650611445b",
      "verified": "sourcify"
    }
  }
} as const;